apps/production-pipeline/requirements-server.txt
# 마리오네트 스튜디오 — 서버 의존성
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
websockets>=12.0

# 기존 파이프라인 의존성
google-genai>=0.5.0
